Adding database outline PDF and python requirements txt file
Added python requirements text file, which contains the python packages needed for interacting with the database in Python. This is mainly a libaray known as 'psycopg2', but other requirements may become neccessary as scripting expands. The requirements.txt file will allow us to keep track of the modules everyone is using, so that we don't run into any overlap or errors in the code.