Added database setup script and some basic python database utilities.
The setup_db.py script should setup the actual database when the server becomes fully operational (uses the database name 'fakenews'). The db_utils.py script is a utility script used by the setup_db.py script. It will also be useful when creating other scripts to interact with the database as it has some nice homebuilt utilities for database manipulation and connection.