Added some error printing for bad connections and moved the reading of the .env file to settings.py
Also added .bat files to .gitignore due to the convience of activating virtual environments from .bat files. A new script has been added to Utils (settings.py) which contains an init function that should be run at the begining of every script in order to load the required keys into the environment.