Connection wrapper changes to allow for a primary connection
The settings.py file now creates a global primary connection and cursor so that connections and cursors don't have to be passed everywhere. A new connection_utils.py file now has most of the previous connection wrappers which take advantage of the new primary connection by defaulting to it. The file also contains various helpful function for executing database queries and commands on the primary connection or on a passed in connection. The db_utils.py file has been updated to take this connection change into account and default to the primary connection by defaulting the connections being passed to None. Many other files still need to be updated to this new primary connection. Specifically the edit and get fakenews_db.py files are still out of date.