Adding some common utilities previously created
Added two common utility functions that will be helpful when populating data. The first is a time_it function which takes in a function and its arguments, then times the executation of that function using the python timeit module and returns the runtime of that function. The second utility is a progress bar function, which takes in various progress measurements and displays a progress bar with progress and ETA when given information about the progress. The progress bar can make use of the time_it function to determine the possible ETA of a loop.