Updated main file for better display and retrieval of data
Mainly updated file so that data could be pulled from an arbitrary number of sources using given specifications. To this end, I added a get_data function which takes in various components that we originally statically definied within the main function. Along with this update, I also updated the printing and statistics so that multiple sources (not just 2) could be handled. Also added a nltk_bayes method which tries to use nltk's native bayesian classifier. Currently this does not work, as the program runs out of memory before it can create the needed classification data.