Added a function for building nltk training data and changed other functions slightly
Specifically, the build_source_dict function was changed so that it returned and built a list of all the words used by a source. This is because nltk may need this type of data for its classifier (and it also may be useful later). Changed other functions which used build_source_dict so they account for the new return.