Fixed zero count error when classifying articles
The classify_article function now takes an extra 'overfit' argument which determines whether, when a source does not contain a given token, you add 1 to the overall score. If overfit is turned on and, when iterating across the tokens in an article, a token is not contained within a source dictionary, the overall score of that source will be increased by 1 (this decreases the similarity, as the lowest score it picked in the end). If overfit is not turned on, tokens that are not in ALL source dictionaries are ignored, so that they do not count towards an article's overall score.